Handled no data page for empty my blog posts
