+++ app/javascript/src/components/common/PageLoader.jsx
@@ -3,7 +3,7 @@ import React from "react";
 import { Spinner } from "neetoui";
 
 const PageLoader = () => (
-  <div className="mt-20 flex w-full items-center justify-center">
+  <div className="mt-20 flex h-80 w-full items-center justify-center">
     <Spinner />
   </div>
 );
